+++ BHoM/Humans/BodyParts/Head.cs
@@ -1,17 +1,17 @@
-﻿using BH.oM.Geometry;
-using BH.oM.Humans.Interfaces;
+﻿//using BH.oM.Geometry;
+//using BH.oM.Humans.Interfaces;
 
-namespace BH.oM.Humans.BodyParts
-{
-    public class Head : IPointBodyPart
-    {
-        /***************************************************/
-        /**** Properties                                ****/
-        /***************************************************/
+//namespace BH.oM.Humans.BodyParts
+//{
+//    public class Head : IPointBodyPart
+//    {
+//        /***************************************************/
+//        /**** Properties                                ****/
+//        /***************************************************/
 
-        public Point TrackingPoint { get; set; } = new Point(); // Point at centre of head
+//        public Point TrackingPoint { get; set; } = new Point(); // Point at centre of head
 
 
-        /***************************************************/
-    }
-}
+//        /***************************************************/
+//    }
+//}
