+++ BHoM/Humans/BodyParts/RightHand.cs
@@ -1,22 +1,22 @@
-﻿using BH.oM.Geometry;
-using BH.oM.Humans.Interfaces;
+﻿//using BH.oM.Geometry;
+//using BH.oM.Humans.Interfaces;
 
-namespace BH.oM.Humans.BodyParts
-{
-    public class RightHand : ILineBodyPart, IPointBodyPart
-    {
-        /***************************************************/
-        /**** Properties                                ****/
-        /***************************************************/
+//namespace BH.oM.Humans.BodyParts
+//{
+//    public class RightHand : ILineBodyPart, IPointBodyPart
+//    {
+//        /***************************************************/
+//        /**** Properties                                ****/
+//        /***************************************************/
 
-        public Point TrackingPoint { get; set; } = new Point();
+//        public Point TrackingPoint { get; set; } = new Point();
 
-        public Line TrackingLine { get; set; } = new Line();
+//        public Line TrackingLine { get; set; } = new Line();
 
-        public HandStateName State { get; set; } = HandStateName.Unknown;
+//        public HandStateName State { get; set; } = HandStateName.Unknown;
 
 
-        /***************************************************/
-    }
+//        /***************************************************/
+//    }
 
-}
+//}
