+++ BHoM/Humans/HandStateName.cs
@@ -1,11 +1,11 @@
-﻿namespace BH.oM.Humans
-{
-    public enum HandStateName
-    {
-        Closed = 0,
-        Lasso,
-        NotTracked,
-        Open,
-        Unknown
-    }
-}
+﻿//namespace BH.oM.Humans
+//{
+//    public enum HandStateName
+//    {
+//        Closed = 0,
+//        Lasso,
+//        NotTracked,
+//        Open,
+//        Unknown
+//    }
+//}
