+++ BHoM/Humans/Interfaces/IPointBodyPart.cs
@@ -1,15 +1,15 @@
-﻿using BH.oM.Base;
-using BH.oM.Geometry;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿//using BH.oM.Base;
+//using BH.oM.Geometry;
+//using System;
+//using System.Collections.Generic;
+//using System.Linq;
+//using System.Text;
+//using System.Threading.Tasks;
 
-namespace BH.oM.Humans.Interfaces
-{
-    interface IPointBodyPart : IObject
-    {
-        Point TrackingPoint { get; set; }
-    }
-}
+//namespace BH.oM.Humans.Interfaces
+//{
+//    interface IPointBodyPart : IObject
+//    {
+//        Point TrackingPoint { get; set; }
+//    }
+//}
