+++ BHoM/Humans/JointName.cs
@@ -1,31 +1,31 @@
-﻿namespace BH.oM.Humans
-{
-    public enum JointName
-    {
-        Head = 0,
-        SpineShoulder, 
-        RightShoulder,
-        LeftShoulder,
-        Neck,
-        ElbowRight,
-        ElbowLeft,
-        RightHand,
-        LeftHand,
-        HandTipRight,
-        HandTipLeft,
-        RightThumb,
-        LeftThumb,
-        WristRight,
-        WristLeft,
-        SpineMid,
-        SpineBase,
-        RightHip,
-        LeftHip,
-        KneeRight,
-        KneeLeft,
-        AnkleRight,
-        AnkleLeft,
-        RightFoot,
-        LeftFoot
-    }
-}
+﻿//namespace BH.oM.Humans
+//{
+//    public enum JointName
+//    {
+//        Head = 0,
+//        SpineShoulder, 
+//        RightShoulder,
+//        LeftShoulder,
+//        Neck,
+//        ElbowRight,
+//        ElbowLeft,
+//        RightHand,
+//        LeftHand,
+//        HandTipRight,
+//        HandTipLeft,
+//        RightThumb,
+//        LeftThumb,
+//        WristRight,
+//        WristLeft,
+//        SpineMid,
+//        SpineBase,
+//        RightHip,
+//        LeftHip,
+//        KneeRight,
+//        KneeLeft,
+//        AnkleRight,
+//        AnkleLeft,
+//        RightFoot,
+//        LeftFoot
+//    }
+//}
