+++ Humans_oM/BodyParts/Head.cs
@@ -1,0 +1,17 @@
+﻿using BH.oM.Geometry;
+using BH.oM.Humans.Interfaces;
+
+namespace BH.oM.Humans.BodyParts
+{
+    public class Head : IPointBodyPart
+    {
+        /***************************************************/
+        /**** Properties                                ****/
+        /***************************************************/
+
+        public Point TrackingPoint { get; set; } = new Point(); // Point at centre of head
+
+
+        /***************************************************/
+    }
+}
