+++ Humans_oM/BodyParts/LeftHand.cs
@@ -1,0 +1,23 @@
+﻿using BH.oM.Base;
+using BH.oM.Geometry;
+using BH.oM.Humans.Interfaces;
+
+namespace BH.oM.Humans.BodyParts
+{
+    public class LeftHand : ILineBodyPart, IPointBodyPart
+    {
+        /***************************************************/
+        /**** Properties                                ****/
+        /***************************************************/
+
+        public Point TrackingPoint { get; set; } = new Point();
+
+        public Line TrackingLine { get; set; } = new Line();
+
+        public HandStateName State { get; set; } = HandStateName.Unknown;
+
+
+        /***************************************************/
+    }
+
+}
