+++ Humans_oM/BodyParts/LeftUpperArm.cs
@@ -1,0 +1,18 @@
+﻿using BH.oM.Geometry;
+using BH.oM.Humans.Interfaces;
+
+namespace BH.oM.Humans.BodyParts
+{
+    public class LeftUpperArm : ILineBodyPart
+    {
+        /***************************************************/
+        /**** Properties                                ****/
+        /***************************************************/
+
+        public Line TrackingLine { get; set; } = new Line();
+
+
+        /***************************************************/
+    }
+
+}
