+++ Humans_oM/BodyParts/RightLowerArm.cs
@@ -1,0 +1,18 @@
+﻿using BH.oM.Geometry;
+using BH.oM.Humans.Interfaces;
+
+namespace BH.oM.Humans.BodyParts
+{
+    public class RightLowerArm : ILineBodyPart
+    {
+        /***************************************************/
+        /**** Properties                                ****/
+        /***************************************************/
+
+        public Line TrackingLine { get; set; } = new Line();
+
+
+        /***************************************************/
+    }
+
+}
