+++ Humans_oM/BodyParts/RightThumb.cs
@@ -1,0 +1,20 @@
+﻿using BH.oM.Geometry;
+using BH.oM.Humans.Interfaces;
+
+namespace BH.oM.Humans.BodyParts
+{
+    public class RightThumb : ILineBodyPart, IPointBodyPart
+    {
+        /***************************************************/
+        /**** Properties                                ****/
+        /***************************************************/
+
+        public Point TrackingPoint { get; set; } = new Point();
+
+        public Line TrackingLine { get; set; } = new Line();
+
+
+        /***************************************************/
+    }
+
+}
