+++ Humans_oM/Enums/HandStateName.cs
@@ -1,0 +1,11 @@
+﻿namespace BH.oM.Humans
+{
+    public enum HandStateName
+    {
+        Closed = 0,
+        Lasso,
+        NotTracked,
+        Open,
+        Unknown
+    }
+}
