+++ Humans_oM/Humans_oM.csproj
@@ -41,7 +41,36 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="BodyParts\Head.cs" />
+    <Compile Include="BodyParts\LeftFoot.cs" />
+    <Compile Include="BodyParts\LeftHand.cs" />
+    <Compile Include="BodyParts\LeftHip.cs" />
+    <Compile Include="BodyParts\LeftLowerArm.cs" />
+    <Compile Include="BodyParts\LeftLowerLeg.cs" />
+    <Compile Include="BodyParts\LeftShoulder.cs" />
+    <Compile Include="BodyParts\LeftThumb.cs" />
+    <Compile Include="BodyParts\LeftUpperArm.cs" />
+    <Compile Include="BodyParts\LeftUpperLeg.cs" />
+    <Compile Include="BodyParts\Neck.cs" />
+    <Compile Include="BodyParts\RightFoot.cs" />
+    <Compile Include="BodyParts\RightHand.cs" />
+    <Compile Include="BodyParts\RightHip.cs" />
+    <Compile Include="BodyParts\RightLowerArm.cs" />
+    <Compile Include="BodyParts\RightLowerLeg.cs" />
+    <Compile Include="BodyParts\RightShoulder.cs" />
+    <Compile Include="BodyParts\RightThumb.cs" />
+    <Compile Include="BodyParts\RightUpperArm.cs" />
+    <Compile Include="BodyParts\RightUpperLeg.cs" />
+    <Compile Include="BodyParts\Spine.cs" />
+    <Compile Include="Enums\HandStateName.cs" />
+    <Compile Include="Employee.cs" />
+    <Compile Include="Human.cs" />
+    <Compile Include="Interfaces\IHumanRole.cs" />
+    <Compile Include="Interfaces\ILineBodyPart.cs" />
+    <Compile Include="Interfaces\IPointBodyPart.cs" />
+    <Compile Include="Enums\JointName.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="Skeleton.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\BHoM\BHoM.csproj">
