+++ Humans_oM/Interfaces/IHumanRole.cs
@@ -1,0 +1,11 @@
+﻿using BH.oM.Base;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace BH.oM.Humans
+{
+    public interface IHumanRole : IObject { }
+}
