+++ Humans_oM/Interfaces/IPointBodyPart.cs
@@ -1,0 +1,15 @@
+﻿using BH.oM.Base;
+using BH.oM.Geometry;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace BH.oM.Humans.Interfaces
+{
+    interface IPointBodyPart : IObject
+    {
+        Point TrackingPoint { get; set; }
+    }
+}
